+++ src/components/MovieReviews.js
@@ -1,7 +1,9 @@
 import { useState, useEffect } from 'react'
-import { Card, Button } from 'react-bootstrap'
+
 import axios from 'axios'
-import SharingButtons from './Sharing/SharingButtons'
+
+import PostCardMovies from './PostCardMovies'
+import { useAuth } from '../context/AuthContext'
 
 import moment from 'moment'
 
@@ -10,6 +12,7 @@ const nytMovieReviewsUrl = `https://api.nytimes.com/svc/movies/v2/reviews/picks.
 
 export default function MovieReviews() {
   const [reviews, setReviews] = useState([])
+  const { currentUser } = useAuth()
   useEffect(() => {
     /* movie reviews */
     axios
@@ -27,38 +30,7 @@ export default function MovieReviews() {
     <div>
       <div className='movie-reviews'>
         {reviews.map((post, index) => (
-          <Card
-            key={index}
-            className='movie-reviews__card  card bg-dark text-light border-light'
-          >
-            <Card.Body>
-              {' '}
-              <div className='title-card'>{post.display_title}</div>
-              <Card.Img
-                className='img'
-                src={post.multimedia.src}
-                alt={post.headline}
-              />
-              <div className='subtitle'>{post.summary_short}</div>
-              <Card.Text className='author-date'>
-                <span>By: {post.byline}</span>{' '}
-                <span>
-                  Published on:{' '}
-                  <span>
-                    {moment(post.publication_date).format('MMMM d, YYYY')}
-                  </span>
-                </span>
-              </Card.Text>
-              <Button className='btn read-more' variant='btn btn-outline-light'>
-                <a href={post.link.url} className='link'>
-                  read more
-                </a>
-              </Button>
-            </Card.Body>
-            <Card.Footer>
-              <SharingButtons url={post.link.url} />
-            </Card.Footer>
-          </Card>
+          <PostCardMovies post={post} user={currentUser} />
         ))}
       </div>
     </div>