+++ src/components/SignUp/SignUp.js
@@ -4,6 +4,7 @@ import { useAuth } from '../../context/AuthContext'
 import { Link, useNavigate } from 'react-router-dom'
 
 export default function SignUp() {
+  const displayNameRef = useRef()
   const emailRef = useRef()
   const passwordRef = useRef()
   const confirmPasswordRef = useRef()
@@ -23,7 +24,11 @@ export default function SignUp() {
     try {
       setError('')
       setLoading(true)
-      await signup(emailRef.current.value, passwordRef.current.value)
+      await signup(
+        emailRef.current.value,
+        passwordRef.current.value,
+        displayNameRef.current.value
+      )
       console.log('account created!')
       navigate('/kiosk')
     } catch {
@@ -43,15 +48,19 @@ export default function SignUp() {
           {/* firebase uses localstorage; also an initial loading state */}
 
           <Form onSubmit={handleSubmit}>
-            <Form.Group id='email'>
+            <Form.Group id='displayName' className='m-2'>
+              <Form.Label>Display name</Form.Label>
+              <Form.Control type='displayName' ref={displayNameRef} required />
+            </Form.Group>
+            <Form.Group id='email' className='m-2'>
               <Form.Label>Email address</Form.Label>
               <Form.Control type='email' ref={emailRef} required />
             </Form.Group>
-            <Form.Group id='password'>
+            <Form.Group id='password' className='m-2'>
               <Form.Label>Password</Form.Label>
               <Form.Control type='password' ref={passwordRef} required />
             </Form.Group>
-            <Form.Group id='confirmPassword'>
+            <Form.Group id='confirmPassword' className='m-2'>
               <Form.Label>Confirm Password</Form.Label>
               <Form.Control type='password' ref={confirmPasswordRef} required />
             </Form.Group>
@@ -59,8 +68,11 @@ export default function SignUp() {
               sign up
             </Button>
           </Form>
-          <div>
-            Already have an account? <Link to='/'>Let's go.</Link>
+          <div className='m-2'>
+            Already have an account?{' '}
+            <Link className='myLink' to='/'>
+              Let's go.
+            </Link>
           </div>
         </Card>
       </div>