+++ src/components/TopStories/BusinessTopStories.js
@@ -3,6 +3,8 @@ import { Card, Button } from 'react-bootstrap'
 import axios from 'axios'
 
 import moment from 'moment'
+import { useAuth } from '../../context/AuthContext'
+import PostCard from '../PostCard'
 
 import SharingButtons from '../Sharing/SharingButtons'
 
@@ -13,6 +15,7 @@ const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/business.jso
 
 export default function BusinnessTopStories() {
   const [posts, setPosts] = useState([])
+  const { currentUser } = useAuth()
   useEffect(() => {
     /* top stories */
     axios
@@ -30,36 +33,9 @@ export default function BusinnessTopStories() {
     <div>
       <div className='top-stories'>
         {posts.map((post, index) => (
-          <Card
-            className='top-stories__card  card bg-dark text-light border-light'
-            key={index}
-          >
-            <Card.Body>
-              {' '}
-              <div className='title-card'>{post.title}</div>
-              <Card.Img
-                className='img'
-                src={post.multimedia[0].url}
-                alt={post.caption}
-              />
-              <div className='subtitle'>{post.abstract}</div>
-              <Card.Text className='author-date'>
-                <span>{post.byline}</span>{' '}
-                <span>
-                  Published:{' '}
-                  {moment(post.published_date).format('MMMM d, YYYY')}
-                </span>
-              </Card.Text>
-              <Button className='btn read-more' variant='btn btn-outline-light'>
-                <a href={post.url} className='link'>
-                  read more
-                </a>
-              </Button>
-            </Card.Body>
-            <Card.Footer>
-              <SharingButtons url={post.url} />
-            </Card.Footer>
-          </Card>
+          <>
+            <PostCard post={post} user={currentUser} />
+          </>
         ))}
       </div>
     </div>