+++ src/components/TopStories/WorldTopStories.js
@@ -1,9 +1,8 @@
 import { useState, useEffect } from 'react'
-import { Card, Button } from 'react-bootstrap'
-import axios from 'axios'
-import SharingButtons from '../Sharing/SharingButtons'
 
-import moment from 'moment'
+import axios from 'axios'
+import PostCard from '../PostCard'
+import { useAuth } from '../../context/AuthContext'
 
 import data from '../../data/nytWorld.json'
 const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/world.json?api-key=${process.env.REACT_APP_NYT_API_KEY}`
@@ -11,13 +10,14 @@ const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/world.json?a
 
 export default function WorldTopStories() {
   const [posts, setPosts] = useState([])
+  const { currentUser } = useAuth()
   useEffect(() => {
     /* top stories */
     axios
       .get(nytTopStoriesUrl)
       .then((response) => {
         console.log(response.data.results)
-        setPosts(response.data.results)
+        setPosts(response.data.results.slice(1))
       })
       .catch((err) => {
         console.log(err)
@@ -28,36 +28,9 @@ export default function WorldTopStories() {
     <div>
       <div className='top-stories'>
         {posts.map((post, index) => (
-          <Card
-            className='top-stories__card  card bg-dark text-light border-light'
-            key={index}
-          >
-            <Card.Body>
-              {' '}
-              <div className='title-card'>{post.title}</div>
-              <Card.Img
-                className='img'
-                src={post.multimedia[0].url}
-                alt={post.caption}
-              />
-              <div className='subtitle'>{post.abstract}</div>
-              <Card.Text className='author-date'>
-                <span>{post.byline}</span>{' '}
-                <span>
-                  Published:{' '}
-                  {moment(post.published_date).format('MMMM d, YYYY')}
-                </span>
-              </Card.Text>
-              <Button className='btn read-more' variant='btn btn-outline-light'>
-                <a href={post.url} className='link'>
-                  read more
-                </a>
-              </Button>
-            </Card.Body>
-            <Card.Footer>
-              <SharingButtons url={post.url} />
-            </Card.Footer>
-          </Card>
+          <>
+            <PostCard post={post} user={currentUser} />
+          </>
         ))}
       </div>
     </div>